Nuke old Annotate code and move over necessary parts
