--- lib/annotate/version.rb
@@ -1,5 +1,0 @@
-module Annotate
-  def self.version
-    '4.0.0'
-  end
-end
