+++ lib/annotate_rb/active_record_patch.rb
