+++ spec/lib/annotate_rb/annotate_models/annotate_models_annotate_model_file_spec.rb
@@ -1,20 +1,15 @@
 # encoding: utf-8
-require 'annotate/annotate_models'
-require 'annotate/active_record_patch'
-require 'active_support/core_ext/string'
-require 'files'
-require 'tmpdir'
 
-RSpec.describe AnnotateModels do
+RSpec.describe AnnotateRb::ModelAnnotator::Annotator do
   describe '.annotate_model_file' do
     before do
       class Foo < ActiveRecord::Base; end
-      allow(AnnotateModels).to receive(:get_model_class).with('foo.rb') { Foo }
+      allow(described_class).to receive(:get_model_class).with('foo.rb') { Foo }
       allow(Foo).to receive(:table_exists?) { false }
     end
 
     subject do
-      AnnotateModels.annotate_model_file([], 'foo.rb', nil, {})
+      described_class.annotate_model_file([], 'foo.rb', nil, {})
     end
 
     after { Object.send :remove_const, 'Foo' }
@@ -26,7 +21,7 @@ RSpec.describe AnnotateModels do
     context 'with a non-class' do
       before do
         NotAClass = 'foo'.freeze # rubocop:disable Naming/ConstantName
-        allow(AnnotateModels).to receive(:get_model_class).with('foo.rb') { NotAClass }
+        allow(described_class).to receive(:get_model_class).with('foo.rb') { NotAClass }
       end
 
       after { Object.send :remove_const, 'NotAClass' }
