+++ spec/lib/annotate_rb/annotate_models/annotate_models_annotating_a_file_frozen_spec.rb
@@ -1,11 +1,6 @@
 # encoding: utf-8
-require 'annotate/annotate_models'
-require 'annotate/active_record_patch'
-require 'active_support/core_ext/string'
-require 'files'
-require 'tmpdir'
 
-RSpec.describe AnnotateModels do
+RSpec.describe AnnotateRb::ModelAnnotator::Annotator do
   include AnnotateTestHelpers
 
   describe 'annotating a file' do
@@ -22,8 +17,8 @@ RSpec.describe AnnotateModels do
                             mock_column(:id, :integer),
                             mock_column(:name, :string, limit: 50)
                           ])
-      @schema_info = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info')
-      Annotate::Helpers.reset_options(Annotate::Constants::ALL_ANNOTATE_OPTIONS)
+      @schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info')
+      AnnotateRb::ModelAnnotator::Helper.reset_options(AnnotateRb::ModelAnnotator::Constants::ALL_ANNOTATE_OPTIONS)
     end
 
     # TODO: Check out why this test fails due to test pollution
@@ -34,7 +29,7 @@ RSpec.describe AnnotateModels do
 
       it "should abort with different annotation when frozen: true " do
         annotate_one_file
-        another_schema_info = AnnotateModels::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
+        another_schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
         @schema_info = another_schema_info
 
         expect { annotate_one_file frozen: true }.to raise_error SystemExit, /user.rb needs to be updated, but annotate was run with `--frozen`./
