+++ spec/lib/annotate_rb/annotate_models/annotate_models_annotating_a_file_spec.rb
@@ -1,11 +1,6 @@
 # encoding: utf-8
-require 'annotate/annotate_models'
-require 'annotate/active_record_patch'
-require 'active_support/core_ext/string'
-require 'files'
-require 'tmpdir'
 
-RSpec.describe AnnotateModels do
+RSpec.describe AnnotateRb::ModelAnnotator::Annotator do
   include AnnotateTestHelpers
   include AnnotateTestConstants
 
@@ -23,8 +18,8 @@ RSpec.describe AnnotateModels do
                             mock_column(:id, :integer),
                             mock_column(:name, :string, limit: 50)
                           ])
-      @schema_info = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info')
-      Annotate::Helpers.reset_options(Annotate::Constants::ALL_ANNOTATE_OPTIONS)
+      @schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info')
+      AnnotateRb::ModelAnnotator::Helper.reset_options(AnnotateRb::ModelAnnotator::Constants::ALL_ANNOTATE_OPTIONS)
     end
 
     context "with 'before'" do
@@ -130,7 +125,7 @@ RSpec.describe AnnotateModels do
                                                 'id',
                                                 on_delete: :cascade)
                              ])
-          @schema_info = AnnotateModels::SchemaInfo.generate(klass, '== Schema Info', show_foreign_keys: true)
+          @schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(klass, '== Schema Info', show_foreign_keys: true)
           annotate_one_file
         end
 
@@ -149,7 +144,7 @@ RSpec.describe AnnotateModels do
                                                 'id',
                                                 on_delete: :restrict)
                              ])
-          @schema_info = AnnotateModels::SchemaInfo.generate(klass, '== Schema Info', show_foreign_keys: true)
+          @schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(klass, '== Schema Info', show_foreign_keys: true)
           annotate_one_file
           expect(File.read(@model_file_name)).to eq("#{@schema_info}#{@file_content}")
         end
@@ -159,7 +154,7 @@ RSpec.describe AnnotateModels do
     describe 'with existing annotation => :before' do
       before do
         annotate_one_file position: :before
-        another_schema_info = AnnotateModels::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
+        another_schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
         @schema_info = another_schema_info
       end
 
@@ -182,7 +177,7 @@ RSpec.describe AnnotateModels do
     describe 'with existing annotation => :after' do
       before do
         annotate_one_file position: :after
-        another_schema_info = AnnotateModels::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
+        another_schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(mock_class(:users, :id, [mock_column(:id, :integer)]), '== Schema Info')
         @schema_info = another_schema_info
       end
 
@@ -203,7 +198,7 @@ RSpec.describe AnnotateModels do
     end
 
     it 'should skip columns with option[:ignore_columns] set' do
-      output = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info',
+      output = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info',
                                                    :ignore_columns => '(id|updated_at|created_at)')
       expect(output.match(/id/)).to be_nil
     end
@@ -220,8 +215,8 @@ RSpec.describe AnnotateModels do
                            mock_column(:id, :integer),
                            mock_column(:name, :string, limit: 50)
                          ])
-      schema_info = AnnotateModels::SchemaInfo.generate(klass, '== Schema Info')
-      AnnotateModels.annotate_one_file(model_file_name, schema_info, position: :before)
+      schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(klass, '== Schema Info')
+      described_class.annotate_one_file(model_file_name, schema_info, position: :before)
       expect(File.read(model_file_name)).to eq("#{schema_info}#{file_content}")
     end
 
@@ -250,7 +245,7 @@ RSpec.describe AnnotateModels do
         model_file_name, = write_model 'user.rb', "#{magic_comment}\n#{content}"
 
         annotate_one_file position: :before
-        schema_info = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info')
+        schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info')
 
         expect(File.read(model_file_name)).to eq("#{magic_comment}\n\n#{schema_info}#{content}")
       end
@@ -259,7 +254,7 @@ RSpec.describe AnnotateModels do
     it 'only keeps a single empty line around the annotation (position :before)' do
       content = "class User < ActiveRecord::Base\nend\n"
       AnnotateTestConstants::MAGIC_COMMENTS.each do |magic_comment|
-        schema_info = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info')
+        schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info')
         model_file_name, = write_model 'user.rb', "#{magic_comment}\n\n\n\n#{content}"
 
         annotate_one_file position: :before
@@ -274,7 +269,7 @@ RSpec.describe AnnotateModels do
         model_file_name, = write_model 'user.rb', "#{magic_comment}\n#{content}"
 
         annotate_one_file position: :after
-        schema_info = AnnotateModels::SchemaInfo.generate(@klass, '== Schema Info')
+        schema_info = AnnotateRb::ModelAnnotator::SchemaInfo.generate(@klass, '== Schema Info')
 
         expect(File.read(model_file_name)).to eq("#{magic_comment}\n#{content}\n#{schema_info}")
       end
@@ -282,7 +277,7 @@ RSpec.describe AnnotateModels do
 
     describe "if a file can't be annotated" do
       before do
-        allow(AnnotateModels).to receive(:get_loaded_model_by_path).with('user').and_return(nil)
+        allow(described_class).to receive(:get_loaded_model_by_path).with('user').and_return(nil)
 
         write_model('user.rb', <<~EOS)
           class User < ActiveRecord::Base
@@ -292,23 +287,23 @@ RSpec.describe AnnotateModels do
       end
 
       it 'displays just the error message with trace disabled (default)' do
-        expect { AnnotateModels.do_annotations model_dir: @model_dir, is_rake: true }.to output(a_string_including("Unable to annotate #{@model_dir}/user.rb: oops")).to_stderr
+        expect { described_class.do_annotations model_dir: @model_dir, is_rake: true }.to output(a_string_including("Unable to annotate #{@model_dir}/user.rb: oops")).to_stderr
 
         # TODO: Find another way of testing trace without hardcoding the file name as part of the spec
-        # expect { AnnotateModels.do_annotations model_dir: @model_dir, is_rake: true }.not_to output(a_string_including('/spec/annotate/annotate_models_spec.rb:')).to_stderr
+        # expect { described_class.do_annotations model_dir: @model_dir, is_rake: true }.not_to output(a_string_including('/spec/annotate/annotate_models_spec.rb:')).to_stderr
       end
 
       it 'displays the error message and stacktrace with trace enabled' do
-        expect { AnnotateModels.do_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("Unable to annotate #{@model_dir}/user.rb: oops")).to_stderr
+        expect { described_class.do_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("Unable to annotate #{@model_dir}/user.rb: oops")).to_stderr
 
         # TODO: Find another way of testing trace without hardcoding the file name as part of the spec
-        # expect { AnnotateModels.do_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including('/spec/lib/annotate/annotate_models_spec.rb:')).to_stderr
+        # expect { described_class.do_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including('/spec/lib/annotate/annotate_models_spec.rb:')).to_stderr
       end
     end
 
     describe "if a file can't be deannotated" do
       before do
-        allow(AnnotateModels).to receive(:get_loaded_model_by_path).with('user').and_return(nil)
+        allow(described_class).to receive(:get_loaded_model_by_path).with('user').and_return(nil)
 
         write_model('user.rb', <<~EOS)
           class User < ActiveRecord::Base
@@ -318,13 +313,13 @@ RSpec.describe AnnotateModels do
       end
 
       it 'displays just the error message with trace disabled (default)' do
-        expect { AnnotateModels.remove_annotations model_dir: @model_dir, is_rake: true }.to output(a_string_including("Unable to deannotate #{@model_dir}/user.rb: oops")).to_stderr
-        expect { AnnotateModels.remove_annotations model_dir: @model_dir, is_rake: true }.not_to output(a_string_including("/user.rb:2:in `<class:User>'")).to_stderr
+        expect { described_class.remove_annotations model_dir: @model_dir, is_rake: true }.to output(a_string_including("Unable to deannotate #{@model_dir}/user.rb: oops")).to_stderr
+        expect { described_class.remove_annotations model_dir: @model_dir, is_rake: true }.not_to output(a_string_including("/user.rb:2:in `<class:User>'")).to_stderr
       end
 
       it 'displays the error message and stacktrace with trace enabled' do
-        expect { AnnotateModels.remove_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("Unable to deannotate #{@model_dir}/user.rb: oops")).to_stderr
-        expect { AnnotateModels.remove_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("/user.rb:2:in `<class:User>'")).to_stderr
+        expect { described_class.remove_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("Unable to deannotate #{@model_dir}/user.rb: oops")).to_stderr
+        expect { described_class.remove_annotations model_dir: @model_dir, is_rake: true, trace: true }.to output(a_string_including("/user.rb:2:in `<class:User>'")).to_stderr
       end
     end
   end
