+++ spec/lib/annotate_rb/annotate_models/annotate_models_set_defaults_spec.rb
@@ -1,14 +1,9 @@
 # encoding: utf-8
-require 'annotate/annotate_models'
-require 'annotate/active_record_patch'
-require 'active_support/core_ext/string'
-require 'files'
-require 'tmpdir'
 
-RSpec.describe AnnotateModels do
+RSpec.describe AnnotateRb::ModelAnnotator::Annotator do
   describe '.set_defaults' do
     subject do
-      Annotate::Helpers.true?(ENV['show_complete_foreign_keys'])
+      AnnotateRb::ModelAnnotator::Helper.true?(ENV['show_complete_foreign_keys'])
     end
 
     context 'when default value of "show_complete_foreign_keys" is not set' do
@@ -19,12 +14,12 @@ RSpec.describe AnnotateModels do
 
     context 'when default value of "show_complete_foreign_keys" is set' do
       before do
-        Annotate.instance_variable_set('@has_set_defaults', false)
-        Annotate.set_defaults('show_complete_foreign_keys' => 'true')
+        AnnotateRb::OldAnnotate.instance_variable_set('@has_set_defaults', false)
+        AnnotateRb::OldAnnotate.set_defaults('show_complete_foreign_keys' => 'true')
       end
 
       after do
-        Annotate.instance_variable_set('@has_set_defaults', false)
+        AnnotateRb::OldAnnotate.instance_variable_set('@has_set_defaults', false)
         ENV.delete('show_complete_foreign_keys')
       end
