--- spec/lib/annotate_spec.rb
@@ -1,7 +1,0 @@
-RSpec.describe Annotate do
-  describe '.version' do
-    it 'has version' do
-      expect(Annotate.version).to be_instance_of(String)
-    end
-  end
-end
